Controller Funcionando na pagina veiculos.html

+++ src/main/java/br/com/estacionamento/estacionamentoBecca/model/Movimentacao.java
@@ -12,47 +12,47 @@ public class Movimentacao {
     @GeneratedValue(strategy= GenerationType.AUTO)
     private Long id;
     @NotNull
-    private Integer id_veiculo;
+    private Long id_veiculo;
 
     @NotNull
-    private Integer id_estacionamento;
+    private Long id_estacionamento;
 
     @NotNull
-    private Integer id_tipo_veiculo;
+    private Long id_tipo_veiculo;
 
     @NotNull
-    private Integer total_valor_estacionado_movimentado;
+    private Long total_valor_estacionado_movimentado;
 
 
-    public Integer getId_veiculo() {
+    public Long getId_veiculo() {
         return id_veiculo;
     }
 
-    public void setId_veiculo(Integer id_veiculo) {
+    public void setId_veiculo(Long id_veiculo) {
         this.id_veiculo = id_veiculo;
     }
 
-    public Integer getId_estacionamento() {
+    public Long getId_estacionamento() {
         return id_estacionamento;
     }
 
-    public void setId_estacionamento(Integer id_estacionamento) {
+    public void setId_estacionamento(Long id_estacionamento) {
         this.id_estacionamento = id_estacionamento;
     }
 
-    public Integer getId_tipo_veiculo() {
+    public Long getId_tipo_veiculo() {
         return id_tipo_veiculo;
     }
 
-    public void setId_tipo_veiculo(Integer id_tipo_veiculo) {
+    public void setId_tipo_veiculo(Long id_tipo_veiculo) {
         this.id_tipo_veiculo = id_tipo_veiculo;
     }
 
-    public Integer getTotal_valor_estacionado_movimentado() {
+    public Long getTotal_valor_estacionado_movimentado() {
         return total_valor_estacionado_movimentado;
     }
 
-    public void setTotal_valor_estacionado_movimentado(Integer total_valor_estacionado_movimentado) {
+    public void setTotal_valor_estacionado_movimentado(Long total_valor_estacionado_movimentado) {
         this.total_valor_estacionado_movimentado = total_valor_estacionado_movimentado;
     }
